refactor: 반응형 컬럼 직렬화 메소드 위치변경
